Mac .dmgs generated by the manifest have upper-case grid names again.
Refactored a lot of the command-line options so that it's more self-documenting.  Now we have a --help option.

Created a separate action for creating unpackeds, which is not part of the default actions (yet).  This should speed up the packaging process.  In the process of doing this, I also made sure to enforce the constraint that no actions are performed on the files in the destination directory after the construct method ends.  To accomplish this I simply moved some strip method calls up into the construct method.

Lastly, I added our copyright notice to installer_template.nsi, and fixed a reference to a non-existant variable.  Hopefully this last won't cause spurious conflicts with Callum's work.

Reviewed by Richard.

Also, the tabs were a little screwed up in the python files, so I untabified and all is good.

--HG--
extra : convert_revision : a3f2d265d2d53353ec1c1488c381f7dc92606bc0