DEV-39997: merging all of the recent xui file updates from the viewer2help-1 branch into viewer-2.0.0-3, to avoid divergence.
This involves giving all floaters and side panels a help_topic
attribute, to specify a logical name for that UI element's page in the
help viewer. If you are creating/updating a floater or panel, please
consider adding a help_topic field. For details, see the wiki page on
XUI Documentation:

  https://wiki.lindenlab.com/wiki/XUI_Documentation#help_topic

I'm also merging in some updates to the lastest floater button
artwork to ensure these are 16x16 pixels.

(Note, the new help code is still being developed in viewer2help-1 and
is NOT being merged into 2.0 with this checkin. This is just data files.)

--HG--
extra : convert_revision : fb1d5cd12a73c9772931099a737398310bb11dee