DEV-39168 - bounce back to the current URL (or, the home URL if current URL is "") if the server denies navigation
This refactors some of the bounceBack code into LLVOVolume.
It also changes an important rule:  the edit panel now *will* send the
current URL to the server when you hit "OK".  This change was done so
that if autoplay is on, we make sure the server gets the right data.

--HG--
extra : convert_revision : 5f3f795f44f885d1435958cf0ab2505b005306a1