Fixes for DEV-41791 and DEV-41920.
Reworked some of the autoplay logic in LLViewerMediaImpl to try and make it more consistent when media is unloaded and reloaded by the performance manager.

LLViewerMediaImpl will now keep track of the media state and current time of media that it unloads for performance reasons, and attempt to restore playing or paused media to the same playhead time and state (playing or paused) when it reloads.

Added "done" status that time-based media plugins can use to indicate that they've reached the end of the media and stopped.

Added logging of priority transitions, for use in debugging issues with the peformance manager.

--HG--
extra : rebase_source : ff4822e660b3aa7dc12ce6e95ae09cdcbd270e87
extra : convert_revision : b43247715389fbba9a803b25951443179191a008