EXT-2231: Let users view release notes for a server change.
I've added a hyperlink to the "You just entered a region using a
different server version" notification. Users can click over this link
to view the release notes for that server version.

This notification recently started turning up in the local chat log,
so the above solution works in that case as well as if it remained as
a popup notification panel.

--HG--
extra : convert_revision : 5681c76576c86f6f0cf638cb0c6375d04e0a80f9