FIX DEV-41991: do not allow media settings panel to come up if media data is in flight Review #33
This change marks the current selection "not editable" if
any objects in the selection are currently "in flight" (i.e.
their media data has not been fetched yet, or is in the
process of being fetched).  This involved adding API to
LLMediaDataClient to query whether an object is in the
process of being fetched (i.e. in the queue).  I've added
a unit test for this new API.

--HG--
extra : convert_revision : 4c9b0ca89dcb34b0f57f7b5ae94483e1340c5c46