EXT-2354 menus contain multiple lines not editable in XUI
We're building a pie menu in code, complete with separators for this particular
branch of submenus. This is terrible and ugly and needs to be killed, but
for now I removed the lines that cause the separators to be added where there
are empty pie slices.

Code reviewed by Vir

--HG--
extra : convert_revision : f06c05fe3ade18c024534918c094dfff16b45098