fix of a normal bug (EXT-2448) [BSI] Notification toasts should resize when text is to long to be displayed
A maximum number of characters for textbox/texteditor could be set only through XML.
In case when a corresponding attribute was absent - the maximun number was set to 255, and there were no chance (no function) to change it.
So in this fix such a function was added.

--HG--
branch : product-engine
extra : convert_revision : 4b4a446d6142b6f6388d754683459277edca02f4