DEV-43041 DEV-41648 - disable OK and Apply if the media data is the same as the initial Review #48
This change kills two birds with one stone: every frame, we check to see if the media data is the same as the one we 'initialize' with. If it is, we disable OK and Cancel. Otherwise we enable them. This makes it very difficult to set empty media data, as well as fixing the "apply button doesn't grey out" bug.

--HG--
extra : convert_revision : afc472a80c623ca043bd0bf8890448fb6d5f2c9e