EXT-2765: Added help IDs for tabs in the Places sidebar.
The issue was that the widgets for this panel are not built solely
from XUI. There's code in LLPanelPlaces::changeInventory() that
creates the actual My Inventory and Teleport History panels. Adding
help_topic strings to the relevant XUI files for the LLLandmarksPanel
and LLTeleportHistoryPanel classes fixes this issue.

--HG--
extra : convert_revision : 05bf0914ce79713115b07440b32f09c14efd0800