EXT-332: Hide international billing features until enabled.
The international billing links and labels in the Buy L$ floater are
now hidden by default. They will only be displayed once we get an
estimate from the server that includes an "estimatedLocalCost" token,
i.e., we are connected to a server that has the international billing
functionality enabled.

This is done so that the new functionality in Viewer 2.0 is not
exposed to the user until the international billing features are live
(which is scheduled to happen after the Viewer 2.0 release date).

--HG--
extra : convert_revision : 5b6ebb00d68d58c7f45d6fe34f22ac129a6101ad