EXT-2510 wrong initial hair color
First try at a fix - some weirdness with when we actually should be setting
the mesh color or not. Fix is untested as the build is broken. Committing for
now so we can re-sync with viewer-2. Will be post-reviewed.

--HG--
branch : avatar-pipeline
extra : convert_revision : 99c26b5782bb5aa2bfa0df46f040d742d7464d6c