Enable OK button in media settings panel even if nothing is changed, but have it be a no-op Review #63
This change addresses an issue where the OK button was greyed out even if nothing was different (changed) in the medias settings floater. This originally was implemented so that no "empty media" data would accidentally get added to the object. But that was confusing (the OK button should never be greyed out). This changes it so that apply() will simply be a no-op if the data has not changed, and now the OK button is always enabled.

--HG--
extra : convert_revision : 75b9d5d24ed0d5cacc9310097e375b9eab6233a6