EXT-3123 cannot decline save for newly created wearable
We weren't marking the initially loaded parameters of a
wearable created through the appearance editor as "saved"
so the client thought the wearable was dirty when it wasn't.
Added a call to mark all parameters as saved right before
sending the wearable to the server.

reviewed by Bigpapi

--HG--
extra : convert_revision : 37badc531556c10312a64d2939dc0c90166be378