EXT-3558: Don't escape '#' chars for URLs in text.
Added '#' to the list of chars that are not escaped when escaping a
URL in a text editor/box. Now, clicking on a link with '#' in text
chat will open the URL in your browser with the '#' intact.

--HG--
extra : convert_revision : c93e3b9d28568a9ac2f97bd9d4d7cc1b05d6a0c4