EXT-3429: Don't display inspect slider when inappropriate.
The avatar inspector has a volume slider and mute button. We make
these widgets invisible when it does not makes sense to interact with
them. That is, when the inspector is for your own avatar, or when
voice is not active.

--HG--
extra : convert_revision : e996bd550e2ff28f063b577420e54b8adaa04091