EXT-4678: Don't hyperlink sim URLs in About window.
Added support for specifying a black list of URLs on a per-widget
basis. URLs on this black list will not be hyperlinked in the text
widget. The About dialog adds the sim hostname to this black list.

--HG--
extra : convert_revision : 9b7955fc7f4fb5a65743351a31ab6e9b51dee193