EXT-3144 EXT-4226: Re-employ tentative state for individual items in media settings Review #95
This change undoes some prior change that seems to not allow showing
media data for multiple selection. There was all of this code put in
to support displaying tentative state for media, but code was added to
basically override it. Perhaps the reason was tentative state items
were not ignored on apply...this change does that.

--HG--
extra : convert_revision : dddb40b50424f617b354c614903a97a2fa2a5486