EXT-4801 : hitting trash button for trashed items does not delete item EXT-4957 : add const correctness for LLFolderViewEventListener::isItemRemovable
This checkin affects several files but is a lot less scary than it looks.  It's mostly to add const correctness, and to rename isInTrash to isItemInTrash so that its naming is consistent with isItemRemovable/isItemMovable.
The only functionality change is to disable the trash button when an item is already in the trash.

--HG--
extra : convert_revision : 3ba0782ceca26164fb9f6b0f56e8110ef23d78a8