EXT-5547 : Autopopulation broken due to new web deploy accounts EXT-5673 : Autopopulation: Created outfits copy subfolder contents as well as immediate folder contents EXT-5632 : Autopopulation is including subfolders even if those aren't outfits DEV-46683 : Post-Deployment Cleanup
This is a series of changes to fix autopopulation behavior that was broken due to the new surprise web avatar deploy.  That deploy surfaced a number of serious issues with the original AP code.  I did not write this code and the person who did is no longer here, so I've done my best to fix those issues up.
This is a fairly comprehensive set of changes, but it's necessary given the poor state of the pre-existing code and how many problems it caused the new web avatar deploy.

This new version of the AP code will:
(1) Look for a Library->Clothing->Initial Outfits folder and use that if it exists
(2) Not create outfits out of folders that aren't complete outfits
(3) No longer string match against "More Outfits" in order to ignore outfits
(4) No longer recursively collect folder contents when creating an outfit (i.e. will only look at direct descendents)

--HG--
extra : convert_revision : 58f0d7bd2031561c2a883015b2bca6c56050ec48