EXT-6114: Ignore secondlife:///app/login SLapps if logged in.
If you were already logged in, clicking on these SLapps would crash
your viewer (crash was in experimental reg-in-client code). We now
ignore these SLapps if the user is already logged in, and a one-time
warning is emitted in this case.

--HG--
extra : convert_revision : 639e351dd6014c8d310edc84390c6b6bfe0570ed