Fixed normal bug EXT-6448 (People -> My Friends list has overlapping text at the top covering the first few contacts.)
Unfortunately I was unable to reproduce this bug. But I have some idea about the reason of it.
Decision to show or not the help text based on count of people in Inventory/Calling Cards/Friends/All list. So, if this list is empty even if Resident has some friend (and at least one of them is online) both help text & online list were shown at the same time.

The reason of why Friends/All can be empty in inventory can be clean cachу & slow connection to upload it from the server. But I could not reproduce it.

So, the fix is to check both all friends & online friends lists to prevent overlaping in the future.

Reviewed by Vadim at https://codereview.productengine.com/secondlife/r/118/

--HG--
branch : product-engine
extra : convert_revision : 552df0103ff84d61920ab29e42d1431dcb391421