EXT-6877 : Show avatar-as-cloud as "Loading..." in nametag
If you or others are a cloud, then your nametag will include a "(Loading...)" string.
Note: this required a small bit of code refactor so that you can query if an avatar (yourself or others) is a cloud; this was basically a rename from updateIsFullyLoaded into getIsCloud.

--HG--
extra : convert_revision : e67bee5946463737e7b78ee39cf007ca619a569f