EXT-7255 FIXED Added dragbar into edit outfit panel.
I was going to implement dragbar via icon rendered in LLResizeBar. Layout panel marked as user_resize=true would have resize bar icon next to it.
But there is an issue in layout stack when user_resize is possible only if there are more than one layout panels with user_resize=true. And applied resize icon (dragbar) is rendered where it is not necessary. See EXT-7471 for details.

So, dragbar was implemented as a workaround: added an icon with dragbar into the "add to outfit" panel.
Also layout of the "add to outfit" panel was updated to match spec.

Reviewed by Vadim Savchuk at https://codereview.productengine.com/secondlife/r/441/

--HG--
branch : product-engine
extra : convert_revision : 5d420914a9065929a4a22554110cf17bae3529e0