EXT-8258 ADDITIONAL FIX Enable the 'Trash' button in 'My Inventory' when a selected item gets worn.
This is an additional fix to the one approved in https://codereview.productengine.com/secondlife/r/772/.

Looks like sometimes get_is_item_worn() fails to determine that an item is actually worn.
I suppose it happens when the item is already linked to COF but not yet known to LLAgentWearables.

So I try fixing it by adding a COF links check to get_is_item_worn(), so that an item is considered worn as soon as it's linked to COF.

Reviewed by Nyx at https://codereview.productengine.com/secondlife/r/822/

--HG--
branch : product-engine
extra : convert_revision : 7ecff5c2e0fbbb6a8e5c04040ddcced0b31d7b00