EXT-8589 FIXED Nothing happens if choose 'Invite to Group' from avatar's 3D context menu
The groups-chooser was auto-closing when the IW selection was lost, which in the case of right-clicking on (other) avatars is immediately.  The 'hg blame' is lost in merge-fubarism so I can't track why this behaviour was ever introduced, but I can think or (or find) no reason to autoclose in this case - the groups picker dialog doesn't even use the current selection, let alone care if it's lost.

--HG--
extra : convert_revision : 489208e86f4d906c92b45ab75596349c014ee3da