EXT-8635 FIXED Replaced "no matches found" with "Loading..." in the My Outfits tab.
Set no items text to "Loading..." when creating the wearable items list and to "There are no items in this outfit" when category is complete.
"Outhfit" is there, because wearable items list nay be empty only for them. This string was introduced, because the only situation when there are
no items shown in outfit tab is for empty outfit- because even when applying filter either all items from outfit will be shown or outfit tab won't be shown at all.

Reviewed by Vadim Savchuk at https://codereview.productengine.com/secondlife/r/865/

--HG--
branch : product-engine
extra : convert_revision : af03630eb89af97fbfd8d76f6912665cd97b9587