downgrading voice to version 2.1 (thats what Imprudence uses) - review for mac and win needed
* its way more stable for me
* so far no lack of functionality found
* in LL viewer 2 currently libvivoxoal is a copy of LL libopenal, though not a binary from the source they provide.
Using our openal instead ( which is compiled from https://bitbucket.org/lindenlab/3p-openal/ )
voice fails with the message libvivoxsdk.so: undefined symbol: alcGetMixedBuffer
Since libvivoxoals license is lgpl but we can't provide the source code we can't distribute that binary.
A friendly person from Linden Lab promised (Jan 11) to take care of the issue, but it would take some time. So giving time by
downgrading seems ok for now. If really necessary we should formally ask LL or Vivox or both for the source in the long run.